+++ migrations/20201210200541_create-messages.sql
@@ -1,0 +1,9 @@
+CREATE TABLE messages (
+    message_id   BIGINT    PRIMARY KEY,
+    author_id    BIGINT    NOT NULL,
+    channel_id   BIGINT    NOT NULL,
+    guild_id     BIGINT    NOT NULL,
+    created      TIMESTAMP NOT NULL,
+    content      TEXT      NOT NULL,
+    msg          JSONB     NOT NULL
+)