+++ src/model/sql/guild/mod.rs
@@ -1,4 +1,5 @@
 pub mod guild_config;
 pub mod guild_roles;
 pub mod guild_setting;
+pub mod messages;
 pub mod tags;