+++ src/model/sql/user/user_data.rs
@@ -128,8 +128,7 @@ impl UserData {
     }
 
     pub async fn from_id(ctx: &Context, user_id: UserId) -> Result<Option<UserData>> {
-        let data = ctx.data.read().await;
-        let pool = data.get::<DbPool>().unwrap();
+        let pool = ctx.data.read().await.get::<DbPool>().cloned().unwrap();
 
         from_id_query(&pool, user_id).await
     }
@@ -141,8 +140,7 @@ impl UserData {
     }
 
     pub async fn save(&self, ctx: &Context) -> Result<UserData> {
-        let data = ctx.data.read().await;
-        let pool = data.get::<DbPool>().unwrap();
+        let pool = ctx.data.read().await.get::<DbPool>().cloned().unwrap();
 
         upsert_query(&pool, &self).await
     }