+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -  Add short `t` alias to `tag` commands
+-  Deleted / edited message logging only if message logs are enabled
 
 ### Fixed